data_pipeline/requirements.txt
pandas==2.0.3
numpy==1.25.2
PyYAML==6.0.1
